cv-analyzer-ai-powered-cv-summary-main/requirements.txt
google-generativeai==0.8.4
PyPDF2==3.0.1
python-dotenv==1.0.0